Move id_allocator and parse_display to the protocol crate
