+++ x11rb-protocol/src/connection/mod.rs
@@ -66,7 +66,7 @@ pub struct Connection {
 }
 
 impl Connection {
-    /// Crate a new `Connection`.
+    /// Create a new `Connection`.
     ///
     /// It is assumed that the connection was just established. This means that the next request
     /// that is sent will have sequence number one.
